Move getLoggerInstance in utils
