+++ scripts/node/utilities.js
@@ -5,45 +5,6 @@
 
 const { exec } = require('child_process');
 const { promisify } = require('util');
-const winston = require('winston');
-
-const LOG_LEVEL_CONFIG = {
-  error: 'red',
-  warn: 'yellow',
-  success: 'green',
-  info: 'cyan',
-  verbose: 'dim white',
-  debug: 'gray',
-};
-
-const createLogger = () => {
-  winston.addColors(LOG_LEVEL_CONFIG);
-  const levels = Object.fromEntries(
-    Object.entries(LOG_LEVEL_CONFIG).map(([level], i) => [level, i]),
-  );
-
-  return winston.createLogger({
-    levels,
-    transports: [
-      new winston.transports.Console({
-        level: 'debug',
-        format: winston.format.combine(
-          winston.format.colorize({ all: true }),
-          winston.format.printf(({ message }) => message),
-        ),
-      }),
-    ],
-  });
-};
-
-let logger;
-
-const getLoggerInstance = () => {
-  if (!logger) {
-    logger = createLogger();
-  }
-  return logger;
-};
 
 const executeCommand = async command => {
   const { stdout } = await promisify(exec)(command);
@@ -58,7 +19,6 @@ const executeCommand = async command => {
 const executeAllCommands = async commands => executeCommand(commands.join(' && '));
 
 module.exports = {
-  getLoggerInstance,
   executeCommand,
   executeAllCommands,
 };
